Cross-Validation is sometimes failing because of joblib bug
We use cross-validation to generate a set of initial scores for the model. Sometimes, this cross-validation fails because of a joblib error when running scikit-learning in parallel. One workaround is to set number_jobs = 1 in the model.yml file, but we want to catch this exception so the model because the initial CV step is not absolutely necessary.